// src/shared/api/openapi/models/LeaderboardEntryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lazy.Blog.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LeaderboardEntryResponse
 */
export interface LeaderboardEntryResponse {
    /**
     * 
     * @type {number}
     * @memberof LeaderboardEntryResponse
     */
    rank: number;
    /**
     * 
     * @type {string}
     * @memberof LeaderboardEntryResponse
     */
    userName: string;
    /**
     * 
     * @type {string}
     * @memberof LeaderboardEntryResponse
     */
    avatarUrl: string | null;
    /**
     * 
     * @type {number}
     * @memberof LeaderboardEntryResponse
     */
    bestScore: number;
    /**
     * 
     * @type {number}
     * @memberof LeaderboardEntryResponse
     */
    gamesPlayed: number;
}

/**
 * Check if a given object implements the LeaderboardEntryResponse interface.
 */
export function instanceOfLeaderboardEntryResponse(value: object): value is LeaderboardEntryResponse {
    if (!('rank' in value) || value['rank'] === undefined) return false;
    if (!('userName' in value) || value['userName'] === undefined) return false;
    if (!('avatarUrl' in value) || value['avatarUrl'] === undefined) return false;
    if (!('bestScore' in value) || value['bestScore'] === undefined) return false;
    if (!('gamesPlayed' in value) || value['gamesPlayed'] === undefined) return false;
    return true;
}

export function LeaderboardEntryResponseFromJSON(json: any): LeaderboardEntryResponse {
    return LeaderboardEntryResponseFromJSONTyped(json, false);
}

export function LeaderboardEntryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): LeaderboardEntryResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'rank': json['rank'],
        'userName': json['userName'],
        'avatarUrl': json['avatarUrl'],
        'bestScore': json['bestScore'],
        'gamesPlayed': json['gamesPlayed'],
    };
}

export function LeaderboardEntryResponseToJSON(json: any): LeaderboardEntryResponse {
    return LeaderboardEntryResponseToJSONTyped(json, false);
}

export function LeaderboardEntryResponseToJSONTyped(value?: LeaderboardEntryResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'rank': value['rank'],
        'userName': value['userName'],
        'avatarUrl': value['avatarUrl'],
        'bestScore': value['bestScore'],
        'gamesPlayed': value['gamesPlayed'],
    };
}